+++ mri/operators/utils/__init__.py
@@ -5,17 +5,6 @@
 #  http://www.cecill.info/licences/Licence_CeCILL-B_V1-en.html for details.   #
 # #############################################################################
 
-
-# Fourier Operators
-from ._fourier.cartesian import FFT
-from ._fourier.non_cartesian import NonCartesianFFT, Stacked3DNFFT
-
-# Linear Operators
-from ._linear.wavelet import WaveletUD2, WaveletN
-from ._linear.dictionary import DictionaryLearning
-
-# Gradient operators
-from ._gradient.gradient import GradSelfCalibrationAnalysis, \
-    GradSelfCalibrationSynthesis, GradSynthesis, GradAnalysis
-
-# Proximity Operators
+# utils functions
+from .._fourier.utils import convert_mask_to_locations, \
+    convert_locations_to_mask, normalize_frequency_locations
